separate RPCMethods from server
