+++ explorer/src/server/handleMessage.ts
@@ -1,26 +1,8 @@
-import { saveJobRunTree, fromString } from '../entity/JobRun'
-import { logger } from '../logging'
 import rpcServer from './rpcServer'
-import { getDb } from '../database'
-
-// todo make jsonRPC
-export const legacyErrorResponse = { status: 422 }
+import handleLegacy from './rpcMethods/legacy'
 
 export type messageContext = {
   chainlinkNodeId: number
-}
-
-const handleLegacy = async (json: string, context: messageContext) => {
-  try {
-    const db = await getDb()
-    const jobRun = fromString(json)
-    jobRun.chainlinkNodeId = context.chainlinkNodeId
-    await saveJobRunTree(db, jobRun)
-    return { status: 201 }
-  } catch (e) {
-    logger.error(e)
-    return { status: 422 }
-  }
 }
 
 const handleJSONRCP = async (request: string, context: messageContext) => {
