+++ explorer/src/server/rpcMethods/index.ts
@@ -1,0 +1,5 @@
+import upsertJobRun from './upsertJobRun'
+
+export default {
+  upsertJobRun,
+}
